Organize plugin build support sources
