+++ emlx/c_src/emlx/native_image.hpp
