+++ emlx/c_src/emlx/plugin/tool_wrapper.cpp
