+++ emlx/c_src/emlx/sha256.cpp
@@ -1,4 +1,4 @@
-#include "emlx_sha256.hpp"
+#include "emlx/sha256.hpp"
 
 #include <algorithm>
 #include <cerrno>
