+++ emlx/test/support/native_image_fixture.cpp
@@ -1,4 +1,4 @@
-#include "emlx_native_image.hpp"
+#include "emlx/native_image.hpp"
 
 #include <cerrno>
 #include <cstdio>
