+++ emlx_axon/Makefile
@@ -66,7 +66,7 @@ $(call esc,$(SCAN_DEP_DIR))/%.d: c_src/%.cpp $(HEADERS) $(PLUGIN_INSPECTOR) $(ca
 	@ "$(PLUGIN_INSPECTOR)" object "$(SCAN_OBJ_DIR)/$*.o" || { rm -f "$@" "$(SCAN_OBJ_DIR)/$*.o"; exit 1; }
 
 $(call esc,$(QWEN3_BUILD_ID_HEADER)): $(SCAN_DEPFILES) $(call esc,$(QWEN3_MLX_ID_HEADER)) $(PLUGIN_DEPENDENCY_POLICY) Makefile
-	"$(EMLX_BUILD_ID_TOOL)" plugin-build-id "$(CURDIR)" "$(EMLX_SOURCE_DIR)" "$(MLX_INCLUDE_DIR)" "$(SCAN_BUILD_ID_HEADER)" "$(QWEN3_MLX_ID_HEADER)" "$(EMLX_PLUGIN_INCLUDE_DIR)/emlx/plugin/build_compat.hpp" "$@" "$(QWEN3_BUILD_ID_FILE)" --policy "$(EMLX_SOURCE_DIR)/c_src/tools/emlx_plugin_tool_wrapper.cpp" --policy "$(EMLX_SOURCE_DIR)/c_src/emlx/plugin/toolchain.hpp" --policy "$(CURDIR)/$(PLUGIN_DEPENDENCY_POLICY)" --policy "$(CURDIR)/$(PLUGIN_INSPECTOR)" --policy "$(CURDIR)/Makefile" --flag "compiler-family=$(EMLX_PLUGIN_COMPILER_FAMILY)" --flag "target=$(TARGET_TRIPLE)" --flag "language=c++20" --flag "optimization=O3" --flag "visibility=hidden" --flag "exceptions=enabled" --flag "rtti=enabled" $(PLUGIN_SCAN_ARGS)
+	"$(EMLX_BUILD_ID_TOOL)" plugin-build-id "$(CURDIR)" "$(EMLX_SOURCE_DIR)" "$(MLX_INCLUDE_DIR)" "$(SCAN_BUILD_ID_HEADER)" "$(QWEN3_MLX_ID_HEADER)" "$(EMLX_PLUGIN_INCLUDE_DIR)/emlx/plugin/build_compat.hpp" "$@" "$(QWEN3_BUILD_ID_FILE)" --policy "$(EMLX_SOURCE_DIR)/c_src/emlx/plugin/tool_wrapper.cpp" --policy "$(EMLX_SOURCE_DIR)/c_src/emlx/plugin/toolchain.hpp" --policy "$(CURDIR)/$(PLUGIN_DEPENDENCY_POLICY)" --policy "$(CURDIR)/$(PLUGIN_INSPECTOR)" --policy "$(CURDIR)/Makefile" --flag "compiler-family=$(EMLX_PLUGIN_COMPILER_FAMILY)" --flag "target=$(TARGET_TRIPLE)" --flag "language=c++20" --flag "optimization=O3" --flag "visibility=hidden" --flag "exceptions=enabled" --flag "rtti=enabled" $(PLUGIN_SCAN_ARGS)
 
 $(call esc,$(QWEN3_BUILD_ID_FILE)): $(call esc,$(QWEN3_BUILD_ID_HEADER))
 	@ test -f "$@"
